# 🚀 Better performance for ctx.Attachment