OLD:
```go
Benchmark_Ctx_Attachment-12      2717662       430 ns/op     128 B/op        2 allocs/op
Benchmark_Ctx_Attachment-12      2781739       443 ns/op     128 B/op        2 allocs/op
Benchmark_Ctx_Attachment-12      2628789       436 ns/op     128 B/op        2 allocs/op
```
NEW:
```go
Benchmark_Ctx_Attachment-12      4048429       292 ns/op     208 B/op        3 allocs/op
Benchmark_Ctx_Attachment-12      4125378       292 ns/op     208 B/op        3 allocs/op
Benchmark_Ctx_Attachment-12      4093616       294 ns/op     208 B/op        3 allocs/op
```